pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.11.0,<3.12"
flask = "^3.0.0"
gunicorn = "^21.2.0"
werkzeug = "^3.1.6"
flask-sqlalchemy = "^3.1.1"
python-dotenv = "^1.2.1"
pyyaml = "^6.0.3"
langchain-community = "^0.4.1"
langchain-huggingface = "^1.2.0"
langchain-text-splitters = "^1.1.1"
chromadb = "^1.5.2"
langchain-chroma = "^1.1.0"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"